requirements.txt
# --- Cœur du Système (Jordan) ---
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9  # Obligatoire pour lire les données USSD (Form Data)

# --- Base de Données & Sécurité (Maxime) ---
sqlalchemy==2.0.27
psycopg2-binary==2.9.9   # Driver pour PostgreSQL
pydantic-settings==2.2.1 # Pour gérer les variables d'environnement (.env)
python-dotenv==1.0.1

# --- Communication & Tests (Emmanuel) ---
requests==2.31.0         # Pour appeler l'API SMS
httpx==0.27.0            # Pour les tests asynchrones de l'API00
africastalking==0.0.1

